consolidate decorate.js
