+++ express/blocks/browse-by-collaboration/browse-by-collaboration.js
@@ -3,7 +3,7 @@
 import {
   createTag,
 } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 
 import buildCarousel from '../shared/carousel.js';
