+++ express/blocks/cta-carousel/cta-carousel.js
@@ -1,5 +1,5 @@
 import { createTag, fetchPlaceholders, transformLinkToAnimation } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 
 import buildCarousel from '../shared/carousel.js';
