+++ express/blocks/floating-button/floating-button.js
@@ -1,4 +1,4 @@
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 import {
   createFloatingButton,
   collectFloatingButtonData,
