+++ express/blocks/fullscreen-marquee-desktop/fullscreen-marquee-desktop.js
@@ -4,7 +4,7 @@ import {
   fetchPlaceholders,
   transformLinkToAnimation,
 } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 import { addFreePlanWidget } from '../../scripts/utils/free-plan.js';
 
 function buildContent(content) {
