+++ express/blocks/gen-ai-cards/gen-ai-cards.js
@@ -1,5 +1,5 @@
 import { createTag, fetchPlaceholders } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 
 import buildCarousel from '../shared/carousel.js';
