+++ express/blocks/link-list/link-list.js
@@ -2,7 +2,7 @@ import {
   fetchRelevantRows,
   normalizeHeadings,
 } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 
 import buildCarousel from '../shared/carousel.js';
