+++ express/blocks/marquee-exp-1/marquee-exp-1.js
@@ -6,7 +6,7 @@ import {
   fetchPlaceholders,
   getConfig,
 } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 import BlockMediator from '../../scripts/block-mediator.min.js';
 
 const breakpointConfig = [
