+++ express/blocks/playlist/playlist.js
@@ -1,6 +1,6 @@
 // eslint-disable-next-line import/no-unresolved
 import { createTag, getIconElement } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 import buildCarousel from '../shared/carousel.js';
 import { fetchVideoAnalytics } from '../shared/video.js';
