+++ express/blocks/pricing-summary/pricing-summary.js
@@ -1,4 +1,4 @@
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 import buildCarousel from '../shared/carousel.js';
 import { createTag } from '../../scripts/utils.js';
