+++ express/blocks/pricing-table/pricing-table.js
@@ -1,7 +1,6 @@
 import { createTag, fetchPlaceholders, yieldToMain } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
 import { debounce } from '../../scripts/hofs.js';
-import { decorateButtons } from '../../scripts/utils/decorate.js';
+import { decorateButtons, addTempWrapper } from '../../scripts/utils/decorate.js';
 import {
   buildUrl,
   fetchPlanOnePlans,
