+++ express/blocks/quick-action-hub/quick-action-hub.js
@@ -3,7 +3,7 @@ import {
   transformLinkToAnimation,
 // eslint-disable-next-line import/no-unresolved
 } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 
 export default function decorate($block) {
   addTempWrapper($block, 'quick-action-hub');
