+++ express/blocks/quotes/quotes.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/no-unresolved
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 import { createTag } from '../../scripts/utils.js';
 
 export default function decorate($block) {
