+++ express/blocks/search-marquee/search-marquee.js
@@ -5,7 +5,7 @@ import {
   getMetadata,
   sampleRUM,
 } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 import { buildFreePlanWidget } from '../../scripts/utils/free-plan.js';
 
 import buildCarousel from '../shared/carousel.js';
