+++ express/blocks/tabs-ax/tabs-ax.js
@@ -3,7 +3,7 @@
  * https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/Tab_Role
  */
 import { createTag } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 
 const MILO_EVENTS = { DEFERRED: 'milo:deferred' };
