+++ express/blocks/template-list-ace/template-list-ace.js
@@ -5,7 +5,7 @@ import {
   getIconElement,
   toClassName,
 } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 import BlockMediator from '../../scripts/block-mediator.min.js';
 
 import { renderModalContent, fetchResults, renderResults } from './results-modal.js';
