+++ express/blocks/toggle-bar/toggle-bar.js
@@ -1,5 +1,5 @@
 import { createTag } from '../../scripts/utils.js';
-import { addTempWrapper } from '../../scripts/decorate.js';
+import { addTempWrapper } from '../../scripts/utils/decorate.js';
 
 function decorateButton(block, toggle) {
   const button = createTag('button', { class: 'toggle-bar-button' });
